Move end2end test project into dapm
