+++ DAPM/DAPM.Test.EndToEnd/Models/GetOrganizationTicketResult.cs
@@ -1,3 +1,3 @@
-namespace EndToEndTest.Models;
+namespace DAPM.Test.EndToEnd.Models;
 
 public record GetOrganizationTicketResult(ICollection<Organization> Organizations);
