+++ DAPM/DAPM.Test.EndToEnd/Models/GetOrganizationsTicketResult.cs
@@ -1,3 +1,3 @@
-namespace EndToEndTest.Models;
+namespace DAPM.Test.EndToEnd.Models;
 
 public record GetOrganizationsTicketResult(ICollection<Organization> Organizations);
