+++ DAPM/DAPM.Test.EndToEnd/Models/GetRepositoriesResult.cs
@@ -1,4 +1,4 @@
-namespace EndToEndTest.Models;
+namespace DAPM.Test.EndToEnd.Models;
 
 public record GetRepositoriesResult(
     Repository[] Repositories
