+++ DAPM/DAPM.Test.EndToEnd/Models/Organization.cs
@@ -1,3 +1,3 @@
-namespace EndToEndTest.Models;
+namespace DAPM.Test.EndToEnd.Models;
 
 public record Organization(Guid Id, string Name, string Domain);
