+++ DAPM/DAPM.Test.EndToEnd/Models/PostRepositoryRequest.cs
@@ -1,3 +1,3 @@
-namespace EndToEndTest.Models;
+namespace DAPM.Test.EndToEnd.Models;
 
 public record PostRepositoryRequest(string name);
