+++ DAPM/DAPM.Test.EndToEnd/Models/PostRepositoryResult.cs
@@ -1,4 +1,4 @@
-namespace EndToEndTest.Models;
+namespace DAPM.Test.EndToEnd.Models;
 
 public record PostRepositoryResult(
     ItemIds ItemIds,
