+++ DAPM/DAPM.Test.EndToEnd/Models/Ticket.cs
@@ -1,3 +1,3 @@
-namespace EndToEndTest.Models;
+namespace DAPM.Test.EndToEnd.Models;
 
 public record Ticket<T>(Guid TicketId, TicketStatus Status, string Message, T Result);
