+++ DAPM/DAPM.Test.EndToEnd/Models/TicketResponse.cs
@@ -1,3 +1,3 @@
-namespace EndToEndTest;
+namespace DAPM.Test.EndToEnd.Models;
 
 public record TicketResponse(string RequestName, Guid TicketId);
