--- EndToEndTest/EndToEndTest.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.6.33829.357
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EndToEndTest", "EndToEndTest\EndToEndTest.csproj", "{8A0506A6-97DB-4680-B1F5-CB71544FDC03}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8A0506A6-97DB-4680-B1F5-CB71544FDC03}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8A0506A6-97DB-4680-B1F5-CB71544FDC03}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8A0506A6-97DB-4680-B1F5-CB71544FDC03}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8A0506A6-97DB-4680-B1F5-CB71544FDC03}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {90EC829B-7216-4247-A638-A8E0618BA9AB}
-	EndGlobalSection
-EndGlobal
